split and restructure modules for matrix generation
