+++ pastis/matrix_generation/matrix_building_numerical.py
@@ -1,12 +1,12 @@
 """
 This module contains functions that construct the matrix M for PASTIS *NUMERICALLY FROM THE RESPECTIVE E2E SIMULATOR*
- and saves it.
+and saves it.
 
  Currently supported:
  JWST
  LUVOIR
  HiCAT
- """
+"""
 
 from abc import ABC, abstractmethod
 import os
@@ -23,13 +23,156 @@ import hcipy
 from pastis.config import CONFIG_PASTIS
 import pastis.util as util
 from pastis.e2e_simulators.hicat_imaging import set_up_hicat
-from pastis.e2e_simulators.luvoir_imaging import LuvoirAPLC, LuvoirA_APLC
+from pastis.e2e_simulators.luvoir_imaging import LuvoirAPLC
 import pastis.e2e_simulators.webbpsf_imaging as webbpsf_imaging
 import pastis.plotting as ppl
 
 log = logging.getLogger()
 matplotlib.rc('image', origin='lower')
 matplotlib.rc('pdf', fonttype=42)
+
+
+class PastisMatrix(ABC):
+    instrument = None
+
+    def __init__(self, design=None, initial_path=''):
+
+        # General telescope parameters
+        self.design = design
+        self.nb_seg = CONFIG_PASTIS.getint(self.instrument, 'nb_subapertures')
+        self.seglist = util.get_segment_list(self.instrument)
+        self.wvln = CONFIG_PASTIS.getfloat(self.instrument, 'lambda') * 1e-9  # m
+        self.wfe_aber = CONFIG_PASTIS.getfloat(self.instrument, 'calibration_aberration') * 1e-9  # m
+
+        # Create directory names
+        tel_suffix = f'{self.instrument.lower()}'
+        if self.instrument == 'LUVOIR':
+            if design is None:
+                design = CONFIG_PASTIS.get('LUVOIR', 'coronagraph_design')
+            tel_suffix += f'-{design}'
+        self.overall_dir = util.create_data_path(initial_path, telescope=tel_suffix)
+        os.makedirs(self.overall_dir, exist_ok=True)
+        self.resDir = os.path.join(self.overall_dir, 'matrix_numerical')
+
+        # Create necessary directories if they don't exist yet
+        os.makedirs(self.resDir, exist_ok=True)
+        os.makedirs(os.path.join(self.resDir, 'OTE_images'), exist_ok=True)
+
+        # Set up logger
+        util.setup_pastis_logging(self.resDir, f'pastis_matrix_{tel_suffix}')
+        log.info(f'Building numerical matrix for {tel_suffix}\n')
+
+        # Record some of the defined parameters
+        log.info(f'Instrument: {tel_suffix}')
+        log.info(f'Wavelength: {self.wvln} m')
+        log.info(f'Number of segments: {self.nb_seg}')
+        log.info(f'Segment list: {self.seglist}')
+        log.info(f'wfe_aber: {self.wfe_aber} m')
+
+        # Copy configfile to resulting matrix directory
+        util.copy_config(self.resDir)
+
+    @abstractmethod
+    def calc(self):
+        """ This is the main method that should be called to calculate a PASTIS matrix. """
+
+
+class PastisMatrixIntensities(PastisMatrix):
+    instrument = None
+
+    def __init__(self, design=None, initial_path='', savepsfs=True, saveopds=True):
+        super().__init__(design=design, initial_path=initial_path)
+
+        self.savepsfs = savepsfs
+        self.saveopds = saveopds
+        self.calculate_matrix_pair = None
+
+        os.makedirs(os.path.join(self.resDir, 'psfs'), exist_ok=True)
+        log.info(f'Total number of actuator pairs in {self.instrument} pupil: {len(list(util.segment_pairs_all(self.nb_seg)))}')
+        log.info(
+            f'Non-repeating pairs in {self.instrument} pupil calculated here: {len(list(util.segment_pairs_non_repeating(self.nb_seg)))}')
+
+    def calc(self):
+        start_time = time.time()
+
+        # Calculate coronagraph floor, and normalization factor from direct image
+        self.calculate_ref_image()
+        self.setup_one_pair_function()
+        self.calculate_contrast_matrix()
+        self.calculate_pastis_from_contrast_matrix()
+
+        end_time = time.time()
+        log.info(
+            f'Runtime for PastisMatrixIntensities().calc(): {end_time - start_time}sec = {(end_time - start_time) / 60}min')
+        log.info(f'Data saved to {self.resDir}')
+
+    def calculate_contrast_matrix(self):
+
+        # Figure out how many processes is optimal and create a Pool.
+        # Assume we're the only one on the machine so we can hog all the resources.
+        # We expect numpy to use multithreaded math via the Intel MKL library, so
+        # we check how many threads MKL will use, and create enough processes so
+        # as to use 100% of the CPU cores.
+        # You might think we should divide number of cores by 2 to get physical cores
+        # to account for hyperthreading, however empirical testing on telserv3 shows that
+        # it is slightly more performant on telserv3 to use all logical cores
+        num_cpu = multiprocessing.cpu_count()
+        # try:
+        #     import mkl
+        #     num_core_per_process = mkl.get_max_threads()
+        # except ImportError:
+        #     # typically this is 4, so use that as default
+        #     log.info("Couldn't import MKL; guessing default value of 4 cores per process")
+        #     num_core_per_process = 4
+
+        num_core_per_process = 1  # NOTE: this was changed by Scott Will in HiCAT and makes more sense, somehow
+        num_processes = int(num_cpu // num_core_per_process)
+        log.info(
+            f"Multiprocess PASTIS matrix for {self.instrument} will use {num_processes} processes (with {num_core_per_process} threads per process)")
+
+        # Iterate over all segment pairs via a multiprocess pool
+        mypool = multiprocessing.Pool(num_processes)
+        t_start = time.time()
+        results = mypool.map(self.calculate_matrix_pair,
+                             util.segment_pairs_non_repeating(self.nb_seg))  # this util function returns a generator
+        t_stop = time.time()
+
+        log.info(f"Multiprocess calculation complete in {t_stop - t_start}sec = {(t_stop - t_start) / 60}min")
+
+        # Unscramble results
+        # results is a list of tuples that contain the return from the partial function, in this case: result[i] = (c, (seg1, seg2))
+        self.contrast_matrix = np.zeros([self.nb_seg, self.nb_seg])  # Generate empty matrix
+        for i in range(len(results)):
+            # Fill according entry in the contrast matrix
+            self.contrast_matrix[results[i][1][0], results[i][1][1]] = results[i][0]
+        mypool.close()
+
+        # Save all contrasts to disk, WITHOUT subtraction of coronagraph floor
+        hcipy.write_fits(self.contrast_matrix, os.path.join(self.resDir, 'contrast_matrix.fits'))
+        plt.figure(figsize=(10, 10))
+        plt.imshow(self.contrast_matrix)
+        plt.colorbar()
+        plt.savefig(os.path.join(self.resDir, 'contrast_matrix.pdf'))
+
+    def calculate_pastis_from_contrast_matrix(self):
+
+        # Calculate the PASTIS matrix from the contrast matrix: analytical matrix element calculation and normalization
+        self.matrix_pastis = pastis_from_contrast_matrix(self.contrast_matrix, self.seglist, self.wfe_aber, float(self.contrast_floor))
+
+        # Save matrix to file
+        filename_matrix = f'pastis_matrix'
+        hcipy.write_fits(self.matrix_pastis, os.path.join(self.resDir, filename_matrix + '.fits'))
+        ppl.plot_pastis_matrix(self.matrix_pastis, self.wvln * 1e9, out_dir=self.resDir, save=True)  # convert wavelength to nm
+        log.info(f'PASTIS matrix saved to: {os.path.join(self.resDir, filename_matrix + ".fits")}')
+
+    @abstractmethod
+    def calculate_ref_image(self):
+        """This method needs to create the attributes self.norm, self.contrast_floor and self.coro_simulator."""
+
+    @abstractmethod
+    def setup_one_pair_function(self):
+        """This needs to create an attribute that is the partial function that can calculate the contrast from one
+        aberrated segment/actuator pair. This needs to create self.calculate_matrix_pair."""
 
 
 def calculate_unaberrated_contrast_and_normalization(instrument, design=None, return_coro_simulator=True, save_coro_floor=False, save_psfs=False, outpath=''):
@@ -188,33 +331,6 @@ def _jwst_matrix_one_pair(norm, wfe_aber, resDir, savepsfs, saveopds, segment_pa
     return contrast, segment_pair
 
 
-def _luvoir_matrix_single_mode(number_all_modes, wfe_aber, luvoir_sim, resDir, saveefields, saveopds, mode_no):
-
-    log.info(f'MODE NUMBER: {mode_no}')
-
-    # Apply calibration aberration to used mode
-    all_modes = np.zeros(number_all_modes)
-    all_modes[mode_no] = wfe_aber / 2
-    luvoir_sim.sm.actuators = all_modes
-
-    # Calculate coronagraphic E-field
-    efield_focal_plane, inter = luvoir_sim.calc_psf(return_intermediate='efield')
-
-    if saveefields:
-        fname_real = f'efield_real_mode{mode_no}'
-        hcipy.write_fits(efield_focal_plane.real, os.path.join(resDir, 'efields', fname_real + '.fits'))
-        fname_imag = f'efield_imag_mode{mode_no}'
-        hcipy.write_fits(efield_focal_plane.imag, os.path.join(resDir, 'efields', fname_imag + '.fits'))
-
-    if saveopds:
-        opd_name = f'opd_mode_{mode_no}'
-        plt.clf()
-        hcipy.imshow_field(inter['seg_mirror'].phase, grid=luvoir_sim.aperture.grid, mask=luvoir_sim.aperture, cmap='RdBu')
-        plt.savefig(os.path.join(resDir, 'OTE_images', opd_name + '.pdf'))
-
-    return efield_focal_plane
-
-
 def _luvoir_matrix_one_pair(design, norm, wfe_aber, resDir, savepsfs, saveopds, segment_pair):
     """
     Function to calculate LVUOIR-A mean contrast of one aberrated segment pair; for PastisMatrixIntensities().
@@ -359,36 +475,6 @@ def pastis_from_contrast_matrix(contrast_matrix, seglist, wfe_aber, coro_floor):
     matrix_pastis = util.symmetrize(matrix_pastis_half)
 
     return matrix_pastis
-
-
-def pastis_matrix_from_efields(electric_fields, efield_ref, direct_norm, dh_mask, wfe_aber):
-
-    # Calculate the semi-analytical PASTIS matrix from the individual E-fields
-    matrix_pastis_half = calculate_semi_analytic_pastis_from_efields(electric_fields, efield_ref, direct_norm, dh_mask)
-
-    # Symmetrize the half-PASTIS matrix
-    log.info('Symmetrizing PASTIS matrix')
-    matrix_pastis = util.symmetrize(matrix_pastis_half)
-
-    # Normalize PASTIS matrix by input aberration
-    matrix_pastis /= np.square(wfe_aber * 1e9)
-
-    return matrix_pastis
-
-
-def calculate_semi_analytic_pastis_from_efields(efields, efield_ref, direct_norm, dh_mask):
-
-    # Create empty matrix
-    nb_modes = efields.shape[0]
-    matrix_pastis_half = np.zeros([nb_modes, nb_modes])
-
-    for pair in util.segment_pairs_non_repeating(nb_modes):
-        intensity_im = np.real((efields[pair[0]].electric_field - efield_ref) * np.conj(efields[pair[1]].electric_field - efield_ref))
-        contrast = util.dh_mean(intensity_im / direct_norm, dh_mask)
-        matrix_pastis_half[pair[0], pair[1]] = contrast
-        log.info(f'Calculated contrast for pair {pair[0]}-{pair[1]}: {contrast}')
-
-    return matrix_pastis_half
 
 
 def calculate_semi_analytic_pastis_from_contrast(contrast_matrix, seglist, coro_floor):
@@ -589,235 +675,6 @@ def num_matrix_multiprocess(instrument, design=None, initial_path='', savepsfs=T
     log.info(f'Data saved to {resDir}')
 
     return overall_dir
-
-
-class PastisMatrix(ABC):
-    instrument = None
-
-    def __init__(self, design=None, initial_path=''):
-
-        # General telescope parameters
-        self.design = design
-        self.nb_seg = CONFIG_PASTIS.getint(self.instrument, 'nb_subapertures')
-        self.seglist = util.get_segment_list(self.instrument)
-        self.wvln = CONFIG_PASTIS.getfloat(self.instrument, 'lambda') * 1e-9  # m
-        self.wfe_aber = CONFIG_PASTIS.getfloat(self.instrument, 'calibration_aberration') * 1e-9  # m
-
-        # Create directory names
-        tel_suffix = f'{self.instrument.lower()}'
-        if self.instrument == 'LUVOIR':
-            if design is None:
-                design = CONFIG_PASTIS.get('LUVOIR', 'coronagraph_design')
-            tel_suffix += f'-{design}'
-        self.overall_dir = util.create_data_path(initial_path, telescope=tel_suffix)
-        os.makedirs(self.overall_dir, exist_ok=True)
-        self.resDir = os.path.join(self.overall_dir, 'matrix_numerical')
-
-        # Create necessary directories if they don't exist yet
-        os.makedirs(self.resDir, exist_ok=True)
-        os.makedirs(os.path.join(self.resDir, 'OTE_images'), exist_ok=True)
-
-        # Set up logger
-        util.setup_pastis_logging(self.resDir, f'pastis_matrix_{tel_suffix}')
-        log.info(f'Building numerical matrix for {tel_suffix}\n')
-
-        # Record some of the defined parameters
-        log.info(f'Instrument: {tel_suffix}')
-        log.info(f'Wavelength: {self.wvln} m')
-        log.info(f'Number of segments: {self.nb_seg}')
-        log.info(f'Segment list: {self.seglist}')
-        log.info(f'wfe_aber: {self.wfe_aber} m')
-
-        # Copy configfile to resulting matrix directory
-        util.copy_config(self.resDir)
-
-    @abstractmethod
-    def calc(self):
-        """ This is the main method that should be called to calculate a PASTIS matrix. """
-
-
-class PastisMatrixIntensities(PastisMatrix):
-    instrument = None
-
-    def __init__(self, design=None, initial_path='', savepsfs=True, saveopds=True):
-        super().__init__(design=design, initial_path=initial_path)
-
-        self.savepsfs = savepsfs
-        self.saveopds = saveopds
-        self.calculate_matrix_pair = None
-
-        os.makedirs(os.path.join(self.resDir, 'psfs'), exist_ok=True)
-        log.info(f'Total number of actuator pairs in {self.instrument} pupil: {len(list(util.segment_pairs_all(self.nb_seg)))}')
-        log.info(
-            f'Non-repeating pairs in {self.instrument} pupil calculated here: {len(list(util.segment_pairs_non_repeating(self.nb_seg)))}')
-
-    def calc(self):
-        start_time = time.time()
-
-        # Calculate coronagraph floor, and normalization factor from direct image
-        self.calculate_ref_image()
-        self.setup_one_pair_function()
-        self.calculate_contrast_matrix()
-        self.calculate_pastis_from_contrast_matrix()
-
-        end_time = time.time()
-        log.info(
-            f'Runtime for PastisMatrixIntensities().calc(): {end_time - start_time}sec = {(end_time - start_time) / 60}min')
-        log.info(f'Data saved to {self.resDir}')
-
-    def calculate_contrast_matrix(self):
-
-        # Figure out how many processes is optimal and create a Pool.
-        # Assume we're the only one on the machine so we can hog all the resources.
-        # We expect numpy to use multithreaded math via the Intel MKL library, so
-        # we check how many threads MKL will use, and create enough processes so
-        # as to use 100% of the CPU cores.
-        # You might think we should divide number of cores by 2 to get physical cores
-        # to account for hyperthreading, however empirical testing on telserv3 shows that
-        # it is slightly more performant on telserv3 to use all logical cores
-        num_cpu = multiprocessing.cpu_count()
-        # try:
-        #     import mkl
-        #     num_core_per_process = mkl.get_max_threads()
-        # except ImportError:
-        #     # typically this is 4, so use that as default
-        #     log.info("Couldn't import MKL; guessing default value of 4 cores per process")
-        #     num_core_per_process = 4
-
-        num_core_per_process = 1  # NOTE: this was changed by Scott Will in HiCAT and makes more sense, somehow
-        num_processes = int(num_cpu // num_core_per_process)
-        log.info(
-            f"Multiprocess PASTIS matrix for {self.instrument} will use {num_processes} processes (with {num_core_per_process} threads per process)")
-
-        # Iterate over all segment pairs via a multiprocess pool
-        mypool = multiprocessing.Pool(num_processes)
-        t_start = time.time()
-        results = mypool.map(self.calculate_matrix_pair,
-                             util.segment_pairs_non_repeating(self.nb_seg))  # this util function returns a generator
-        t_stop = time.time()
-
-        log.info(f"Multiprocess calculation complete in {t_stop - t_start}sec = {(t_stop - t_start) / 60}min")
-
-        # Unscramble results
-        # results is a list of tuples that contain the return from the partial function, in this case: result[i] = (c, (seg1, seg2))
-        self.contrast_matrix = np.zeros([self.nb_seg, self.nb_seg])  # Generate empty matrix
-        for i in range(len(results)):
-            # Fill according entry in the contrast matrix
-            self.contrast_matrix[results[i][1][0], results[i][1][1]] = results[i][0]
-        mypool.close()
-
-        # Save all contrasts to disk, WITHOUT subtraction of coronagraph floor
-        hcipy.write_fits(self.contrast_matrix, os.path.join(self.resDir, 'contrast_matrix.fits'))
-        plt.figure(figsize=(10, 10))
-        plt.imshow(self.contrast_matrix)
-        plt.colorbar()
-        plt.savefig(os.path.join(self.resDir, 'contrast_matrix.pdf'))
-
-    def calculate_pastis_from_contrast_matrix(self):
-
-        # Calculate the PASTIS matrix from the contrast matrix: analytical matrix element calculation and normalization
-        self.matrix_pastis = pastis_from_contrast_matrix(self.contrast_matrix, self.seglist, self.wfe_aber, float(self.contrast_floor))
-
-        # Save matrix to file
-        filename_matrix = f'pastis_matrix'
-        hcipy.write_fits(self.matrix_pastis, os.path.join(self.resDir, filename_matrix + '.fits'))
-        ppl.plot_pastis_matrix(self.matrix_pastis, self.wvln * 1e9, out_dir=self.resDir, save=True)  # convert wavelength to nm
-        log.info(f'PASTIS matrix saved to: {os.path.join(self.resDir, filename_matrix + ".fits")}')
-
-    @abstractmethod
-    def calculate_ref_image(self):
-        """This method needs to create the attributes self.norm, self.contrast_floor and self.coro_simulator."""
-
-    @abstractmethod
-    def setup_one_pair_function(self):
-        """This needs to create an attribute that is the partial function that can calculate the contrast from one
-        aberrated segment/actuator pair. This needs to create self.calculate_matrix_pair."""
-
-
-class PastisMatrixEfields(PastisMatrix):
-    instrument = None
-
-    def __init__(self, design=None, initial_path='', saveefields=True, saveopds=True):
-        super().__init__(design=design, initial_path=initial_path)
-
-        self.save_efields = saveefields
-        self.saveopds = saveopds
-        self.calculate_one_mode = None
-        self.efields_per_mode = []
-
-        os.makedirs(os.path.join(self.resDir, 'efields'), exist_ok=True)
-
-    def calc(self):
-        start_time = time.time()
-
-        self.calculate_ref_efield()
-        self.setup_deformable_mirror()
-        self.setup_single_mode_function()
-        self.calculate_efields()
-        self.calculate_pastis_matrix_from_efields()
-
-        end_time = time.time()
-        log.info(
-            f'Runtime for PastisMatrixEfields().calc(): {end_time - start_time}sec = {(end_time - start_time) / 60}min')
-        log.info(f'Data saved to {self.resDir}')
-
-    def calculate_efields(self):
-        for i in range(self.number_all_modes):
-            self.efields_per_mode.append(self.calculate_one_mode(i))
-        self.efields_per_mode = np.array(self.efields_per_mode)
-
-    def calculate_pastis_matrix_from_efields(self):
-        self.matrix_pastis = pastis_matrix_from_efields(self.efields_per_mode, self.efield_ref, self.norm, self.dh_mask, self.wfe_aber)
-
-        # Save matrix to file
-        filename_matrix = f'pastis_matrix'
-        hcipy.write_fits(self.matrix_pastis, os.path.join(self.resDir, filename_matrix + '.fits'))
-        ppl.plot_pastis_matrix(self.matrix_pastis, self.wvln * 1e9, out_dir=self.resDir, save=True)  # convert wavelength to nm
-        log.info(f'PASTIS matrix saved to: {os.path.join(self.resDir, filename_matrix + ".fits")}')
-
-    @abstractmethod
-    def calculate_ref_efield(self):
-        pass
-
-    @abstractmethod
-    def setup_deformable_mirror(self):
-        pass
-
-    @abstractmethod
-    def setup_single_mode_function(self):
-        pass
-
-
-class MatrixEfieldLuvoirA(PastisMatrixEfields):
-    instrument = 'LUVOIR'
-
-    def __init__(self, design='small', max_local_zernike=3, initial_path='', saveefields=True, saveopds=True):
-        super().__init__(design=design, initial_path=initial_path, saveefields=saveefields, saveopds=saveopds)
-        self.max_local_zernike = max_local_zernike
-
-    def calculate_ref_efield(self):
-        optics_input = os.path.join(util.find_repo_location(), CONFIG_PASTIS.get('LUVOIR', 'optics_path_in_repo'))
-        sampling = CONFIG_PASTIS.getfloat('LUVOIR', 'sampling')
-        self.luvoir = LuvoirA_APLC(optics_input, self.design, sampling)
-        self.dh_mask = self.luvoir.dh_mask
-
-        # Calculate contrast normalization factor from direct PSF (intensity)
-        _unaberrated_coro_psf, direct = self.luvoir.calc_psf(ref=True)
-        self.norm = np.max(direct)
-
-        # Calculate reference E-field in focal plane, without any aberrations applied
-        unaberrated_ref_efield, _inter = self.luvoir.calc_psf(return_intermediate='efield')
-        self.efield_ref = unaberrated_ref_efield.electric_field
-
-    def setup_deformable_mirror(self):
-        log.info(f'Creating segmented mirror with {self.max_local_zernike} local modes each...')
-        self.luvoir.create_segmented_mirror(self.max_local_zernike)
-        self.number_all_modes = self.luvoir.sm.num_actuators
-        log.info(f'Total number of modes: {self.number_all_modes}')
-
-    def setup_single_mode_function(self):
-        self.calculate_one_mode = functools.partial(_luvoir_matrix_single_mode, self.number_all_modes, self.wfe_aber,
-                                                    self.luvoir, self.resDir, self.save_efields, self.saveopds)
 
 
 class MatrixIntensityLuvoirA(PastisMatrixIntensities):
